data-cleansing-sot/solutions/requirements.txt
pandas
psycopg2-binary
python-dotenv
pytest
